chore: move rename action
