+++ routes/web.php
@@ -54,7 +54,7 @@ Route::middleware(['auth', 'verified', 'throttle:60,1', 'set.locale'])->group(fu
 
             // settings
             Route::get('journals/{slug}/settings', [Journals\Settings\JournalSettingsController::class, 'show'])->name('journal.settings.show');
-            Route::put('journals/{slug}/settings', [Journals\Settings\JournalSettingsController::class, 'update'])->name('journal.settings.update');
+            Route::put('journals/{slug}', [Journals\JournalController::class, 'update'])->name('journal.update');
             Route::delete('journals/{slug}', [Journals\JournalController::class, 'destroy'])->name('journal.destroy');
         });
     });
