+++ tests/Feature/Controllers/App/Journals/JournalControllerTest.php
@@ -8,8 +8,8 @@ use App\Models\Journal;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Hash;
-use Tests\TestCase;
 use PHPUnit\Framework\Attributes\Test;
+use Tests\TestCase;
 
 final class JournalControllerTest extends TestCase
 {
@@ -83,6 +83,54 @@ final class JournalControllerTest extends TestCase
         $response = $this->actingAs($user)->get('/journals/' . $journal->slug);
 
         $response->assertStatus(404);
+    }
+
+    #[Test]
+    public function it_updates_the_journal_name(): void
+    {
+        $user = User::factory()->create();
+        $journal = Journal::factory()->for($user)->create([
+            'name' => 'Dunder Mifflin',
+        ]);
+
+        $response = $this->actingAs($user)->put('/journals/' . $journal->slug, [
+            'journal_name' => 'Threat Level Midnight',
+        ]);
+
+        $journal->refresh();
+
+        $response->assertRedirect(route('journal.settings.show', ['slug' => $journal->slug]));
+        $response->assertSessionHas('status', 'Changes saved');
+        $this->assertEquals('Threat Level Midnight', $journal->name);
+    }
+
+    #[Test]
+    public function it_validates_the_journal_name_when_updating(): void
+    {
+        $user = User::factory()->create();
+        $journal = Journal::factory()->for($user)->create();
+
+        $response = $this->actingAs($user)
+            ->from('/journals/' . $journal->slug . '/settings')
+            ->put('/journals/' . $journal->slug, [
+                'journal_name' => 'Invalid@!',
+            ]);
+
+        $response->assertRedirect('/journals/' . $journal->slug . '/settings');
+        $response->assertSessionHasErrors(['journal_name']);
+    }
+
+    #[Test]
+    public function it_returns_not_found_when_trying_to_update_a_journal_the_user_does_not_own(): void
+    {
+        $user = User::factory()->create();
+        $otherJournal = Journal::factory()->create();
+
+        $response = $this->actingAs($user)->put('/journals/' . $otherJournal->slug, [
+            'journal_name' => 'Threat Level Midnight',
+        ]);
+
+        $response->assertNotFound();
     }
 
     #[Test]
